pyproject.toml
[project]
name = "nvim-config"
version = "0.1.0"
description = "Prateek's Neovim config"
authors = [
    { name = "Prateek Raman", email = "vvnraman@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "pydata-sphinx-theme>=0.16.1",
    "sphinx>=8.2.3",
    "sphinx-autobuild>=2025.8.25",
    "sphinx-tabs>=3.4.7",
    "sphinxcontrib-mermaid>=1.0.0",
    "sphinxemoji>=0.3.1",
    "typer-slim>=0.20.0",
]

[project.scripts]
nvim-config = "nvim_config.main:app"

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[tool.basedpyright]
include = ["src", "docs"]
exclude = [
  "**/__pycache__",
  "**/*.egg-info",
  "docs/_build"
]
ignore = [
  "lua/",
]

reportMissingImports = "error"
